+++ common/src/main/java/com/wynntils/core/features/FeatureManager.java
@@ -112,6 +112,7 @@ import com.wynntils.features.tooltips.TooltipVanillaHideFeature;
 import com.wynntils.features.trademarket.TradeMarketAutoOpenChatFeature;
 import com.wynntils.features.trademarket.TradeMarketBulkSellFeature;
 import com.wynntils.features.trademarket.TradeMarketPriceConversionFeature;
+import com.wynntils.features.trademarket.TradeMarketPriceMatchFeature;
 import com.wynntils.features.ui.BulkBuyFeature;
 import com.wynntils.features.ui.ContainerScrollFeature;
 import com.wynntils.features.ui.CosmeticsPreviewFeature;
@@ -121,7 +122,6 @@ import com.wynntils.features.ui.CustomSeaskipperScreenFeature;
 import com.wynntils.features.ui.LobbyUptimeFeature;
 import com.wynntils.features.ui.SoulPointTimerFeature;
 import com.wynntils.features.ui.TradeMarketBulkSellFeature;
-import com.wynntils.features.ui.TradeMarketPriceMatchFeature;
 import com.wynntils.features.ui.WynncraftButtonFeature;
 import com.wynntils.features.ui.WynncraftPauseScreenFeature;
 import com.wynntils.features.ui.WynntilsContentBookFeature;
@@ -289,6 +289,7 @@ public final class FeatureManager extends Manager {
         registerFeature(new TradeMarketAutoOpenChatFeature());
         registerFeature(new TradeMarketBulkSellFeature());
         registerFeature(new TradeMarketPriceConversionFeature());
+        registerFeature(new TradeMarketPriceMatchFeature());
         // endregion
 
         // region ui
@@ -300,8 +301,6 @@ public final class FeatureManager extends Manager {
         registerFeature(new CustomSeaskipperScreenFeature());
         registerFeature(new LobbyUptimeFeature());
         registerFeature(new SoulPointTimerFeature());
-        registerFeature(new TradeMarketBulkSellFeature());
-        registerFeature(new TradeMarketPriceMatchFeature());
         registerFeature(new WynncraftButtonFeature());
         registerFeature(new WynncraftPauseScreenFeature());
         registerFeature(new WynntilsContentBookFeature());
